pdp variants
